// client/src/shared/api/generated/models/dataset-list-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenGraph API
 * AI/ML model and dataset Web3 blockchain infrastructure server
 *
 * The version of the OpenAPI document: 0.1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

// May contain unused imports in some cases
// @ts-ignore
import type { DatasetListItem } from "./dataset-list-item";

/**
 * Dataset list response schema
 * @export
 * @interface DatasetListResponse
 */
export interface DatasetListResponse {
  /**
   * Dataset list
   * @type {Array<DatasetListItem>}
   * @memberof DatasetListResponse
   */
  items: Array<DatasetListItem>;
  /**
   * Total count
   * @type {number}
   * @memberof DatasetListResponse
   */
  total: number;
  /**
   * Current page
   * @type {number}
   * @memberof DatasetListResponse
   */
  page: number;
  /**
   * Page size
   * @type {number}
   * @memberof DatasetListResponse
   */
  limit: number;
  /**
   * Total pages
   * @type {number}
   * @memberof DatasetListResponse
   */
  pages: number;
}