// client/src/shared/api/generated/models/image-list-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenGraph API
 * AI/ML model and dataset Web3 blockchain infrastructure server
 *
 * The version of the OpenAPI document: 0.1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

// May contain unused imports in some cases
// @ts-ignore
import type { ImageRead } from "./image-read";

/**
 * Image list response schema
 * @export
 * @interface ImageListResponse
 */
export interface ImageListResponse {
  /**
   * Image list
   * @type {Array<ImageRead>}
   * @memberof ImageListResponse
   */
  items: Array<ImageRead>;
  /**
   * Total count
   * @type {number}
   * @memberof ImageListResponse
   */
  total: number;
  /**
   * Current page
   * @type {number}
   * @memberof ImageListResponse
   */
  page: number;
  /**
   * Page size
   * @type {number}
   * @memberof ImageListResponse
   */
  limit: number;
  /**
   * Total pages
   * @type {number}
   * @memberof ImageListResponse
   */
  pages: number;
}